+++ test/storage.spec.js
@@ -2,32 +2,38 @@
 
 const express = require('express');
 const chai = require('chai');
-const expect = chai.expect;
-const GridFsStorage = require('../index');
-const setting = require('./utils/settings');
 const request = require('supertest');
 const multer = require('multer');
+const mongoose = require('mongoose');
 const mongo = require('mongodb');
-const MongoClient = mongo.MongoClient;
 const md5File = require('md5-file');
 const fs = require('fs');
-const cache = GridFsStorage.cache;
+const sinon = require('sinon');
+const sinonChai = require('sinon-chai');
+const util = require('util');
+
+const GridFsStorage = require('../index');
+const settings = require('./utils/settings');
 const testUtils = require('./utils/testutils');
+
+const expect = chai.expect;
+const MongoClient = mongo.MongoClient;
 const files = testUtils.files;
 const cleanStorage = testUtils.cleanStorage;
 const getDb = testUtils.getDb;
 const getClient = testUtils.getClient;
+const storageReady = testUtils.storageReady;
+
+chai.use(sinonChai);
 
 describe('Storage', () => {
   let result, app, storage;
 
   before(() => app = express());
 
-  beforeEach(() => cache.clear());
-
   describe('url created instance', () => {
     before((done) => {
-      storage = new GridFsStorage({url: setting.mongoUrl});
+      storage = new GridFsStorage({url: settings.mongoUrl});
 
       const upload = multer({storage});
 
@@ -63,10 +69,10 @@ describe('Storage', () => {
 
   describe('db created instance', () => {
     before(() => {
-      const promise = MongoClient.connect(setting.mongoUrl);
+      const promise = MongoClient.connect(settings.mongoUrl);
       return promise.then(_db => {
         const db = getDb(_db);
-        storage = GridFsStorage({db});
+        storage = new GridFsStorage({db});
         storage.client = getClient(_db);
 
         const upload = multer({storage});
@@ -102,18 +108,80 @@ describe('Storage', () => {
 
   });
 
+  describe('Mongoose', () => {
+    let client, db;
+
+    it('should connect using a mongoose instance', (done) => {
+      const promise = mongoose.connect(settings.mongoUrl, {useNewUrlParser: true});
+      storage = new GridFsStorage({db: promise});
+      const upload = multer({storage});
+
+      app.post('/mongoose_instance', upload.array('photos', 2), (req, res) => {
+        result = {headers: req.headers, files: req.files, body: req.body};
+        res.end();
+      });
+
+      storage.on('connection', (_db) => {
+        expect(_db).to.be.instanceOf(mongoose.mongo.Db);
+        db = _db;
+        client = mongoose.connection;
+
+        request(app)
+          .post('/mongoose_instance')
+          .attach('photos', files[0])
+          .attach('photos', files[1])
+          .end((err) => {
+            expect(err).to.equal(null);
+            expect(result.files.length).to.equal(2);
+            done(err);
+          });
+      });
+    });
+
+    it('should connect using a mongoose connection', (done) => {
+      const promise = mongoose
+        .connect(settings.mongoUrl, {useNewUrlParser: true})
+        .then(instance => instance.connection);
+      storage = new GridFsStorage({db: promise});
+      const upload = multer({storage});
+
+      app.post('/mongoose_connection', upload.array('photos', 2), (req, res) => {
+        result = {headers: req.headers, files: req.files, body: req.body};
+        res.end();
+      });
+
+      storage.on('connection', (_db) => {
+        expect(_db).to.be.instanceOf(mongoose.mongo.Db);
+        db = _db;
+        client = mongoose.connection;
+
+        request(app)
+          .post('/mongoose_connection')
+          .attach('photos', files[0])
+          .attach('photos', files[1])
+          .end((err) => {
+            expect(err).to.equal(null);
+            expect(result.files.length).to.equal(2);
+            done(err);
+          });
+      });
+    });
+
+    afterEach(() => cleanStorage(storage, db, client));
+  });
+
   describe('db promise based instance', () => {
     let db, client;
     before((done) => {
 
       const promised = MongoClient
-        .connect(setting.mongoUrl).then(_db => {
+        .connect(settings.mongoUrl).then(_db => {
           db = getDb(_db);
           client = getClient(_db);
           return db;
         });
 
-      storage = GridFsStorage({db: promised});
+      storage = new GridFsStorage({db: promised});
       const upload = multer({storage});
 
 
@@ -149,7 +217,7 @@ describe('Storage', () => {
     let db, client;
     before((done) => {
       const promised = MongoClient
-        .connect(setting.mongoUrl).then((_db) => {
+        .connect(settings.mongoUrl).then((_db) => {
           db = getDb(_db);
           client = getClient(_db);
           return new Promise((resolve) => {
@@ -157,7 +225,7 @@ describe('Storage', () => {
           });
         });
 
-      storage = GridFsStorage({db: promised});
+      storage = new GridFsStorage({db: promised});
       const upload = multer({storage});
 
 
@@ -191,7 +259,7 @@ describe('Storage', () => {
   describe('default uploaded file spec', () => {
     let size;
     before((done) => {
-      storage = GridFsStorage({url: setting.mongoUrl});
+      storage = new GridFsStorage({url: settings.mongoUrl});
       const upload = multer({storage});
 
       app.post('/spec', upload.single('photo'), (req, res) => {
@@ -254,7 +322,7 @@ describe('Storage', () => {
       const names = ['plants', 'animals'];
       const contentTypes = ['text/plain', 'image/jpeg'];
       storage = new GridFsStorage({
-        url: setting.mongoUrl,
+        url: settings.mongoUrl,
         file: () => {
           counter++;
           return {
@@ -331,8 +399,8 @@ describe('Storage', () => {
 
   describe('Missing properties in file naming function', () => {
     before((done) => {
-      storage = GridFsStorage({
-        url: setting.mongoUrl,
+      storage = new GridFsStorage({
+        url: settings.mongoUrl,
         file: () => {
           return {
             metadata: {foo: 'bar'},
@@ -391,8 +459,8 @@ describe('Storage', () => {
       const values = [null, undefined, {}];
       let counter = -1;
 
-      storage = GridFsStorage({
-        url: setting.mongoUrl,
+      storage = new GridFsStorage({
+        url: settings.mongoUrl,
         file: () => {
           counter++;
           return values[counter];
@@ -447,8 +515,8 @@ describe('Storage', () => {
       const values = ['name', 10];
       let counter = -1;
 
-      storage = GridFsStorage({
-        url: setting.mongoUrl,
+      storage = new GridFsStorage({
+        url: settings.mongoUrl,
         file: () => {
           counter++;
           return values[counter];
@@ -496,6 +564,154 @@ describe('Storage', () => {
     after(() => cleanStorage(storage));
   });
 
+  describe('Using connection options', () => {
+    let storage2;
+
+    beforeEach(() => storage2 = null);
+
+    it('should be compatible with an options object on url based connections', (done) => {
+      storage = new GridFsStorage({
+        url: settings.mongoUrl,
+        options: {
+          poolSize: 10,
+        },
+      });
+
+      storage.on('connection', () => {
+        expect(storage.db.serverConfig.s.poolSize).to.equal(10);
+        done();
+      });
+    });
+
+    it('should preserve compatibility with a connectionOpts options property', (done) => {
+      const version = process.version.slice(1).split('.').map(Number);
+      const warningSpy = sinon.spy();
+      const deprecate = sinon.stub(util, 'deprecate').callThrough();
+      process.on('warning', warningSpy);
+
+      storage = new GridFsStorage({
+        url: settings.mongoUrl,
+        connectionOpts: {
+          poolSize: 10,
+        },
+      });
+
+      storage2 = GridFsStorage({
+        url: settings.mongoUrl,
+        connectionOpts: {
+          poolSize: 10,
+        },
+      });
+
+      Promise
+        .all(storageReady(storage, storage2))
+        .then(() => {
+          const expectedMessage = 'The property "connectionOpts" is deprecated. Use "options" instead.';
+          expect(storage.db.serverConfig.s.poolSize).to.equal(10);
+          expect(deprecate).to.have.callCount(1);
+          expect(deprecate.getCall(0).args[1]).to.equal(expectedMessage);
+          if (version[0] !== 4) {
+            expect(warningSpy).to.have.callCount(1);
+            expect(warningSpy.getCall(0).args[0].message).to.equal(expectedMessage);
+          }
+          done();
+        })
+        .catch(done);
+    });
+
+    afterEach(() => {
+      return Promise.all([
+        cleanStorage(storage),
+        storage2 ? cleanStorage(storage2) : Promise.resolve(),
+      ]);
+    });
+  });
+
+  describe('Ready method', () => {
+    let result, resolveSpy, rejectSpy;
+
+    describe('Connection successful', () => {
+      beforeEach(() => {
+        storage = new GridFsStorage({
+          url: settings.mongoUrl
+        });
+      });
+
+      it('should return a promise that resolves when the connection is created', () => {
+        result = storage.ready();
+        resolveSpy = sinon.spy();
+        rejectSpy = sinon.spy();
+        storage.once('connection', resolveSpy);
+        storage.once('connectionFailed', rejectSpy);
+        expect(result).to.be.a('promise');
+        result.then(db => {
+          expect(resolveSpy).to.have.callCount(1);
+          expect(rejectSpy).to.have.callCount(0);
+          expect(db).to.equal(storage.db);
+          expect(db).not.to.equal(null);
+        });
+        return result;
+      });
+
+      it('should return a promise that resolves if the connection is already created', (done) => {
+        storage.once('connection', () => {
+          result = storage.ready();
+          expect(result).to.be.a('promise');
+          result
+            .then(db => {
+              expect(db).to.equal(storage.db);
+              expect(db).not.to.equal(null);
+              done();
+            })
+            .catch(done);
+        });
+      });
+    });
+
+    describe('Connection failed', () => {
+      const error = new Error('Fake error');
+      beforeEach(() => {
+        sinon.stub(MongoClient, 'connect').callsFake(function () {
+          setTimeout(() => arguments[2](error));
+        });
+        storage = new GridFsStorage({
+          url: settings.mongoUrl
+        });
+      });
+
+      it('should return a promise that rejects when the connection fails', () => {
+        result = storage.ready();
+        resolveSpy = sinon.spy();
+        rejectSpy = sinon.spy();
+        storage.once('connection', resolveSpy);
+        storage.once('connectionFailed', rejectSpy);
+
+        expect(result).to.be.a('promise');
+        return result.catch(err => {
+          expect(resolveSpy).to.have.callCount(0);
+          expect(rejectSpy).to.have.callCount(1);
+          expect(err).to.equal(rejectSpy.getCall(0).args[0]);
+          expect(err).to.equal(error);
+        });
+      });
+
+      it('should return a promise that rejects if the module already failed connecting', (done) => {
+        storage.once('connectionFailed', (evtErr) => {
+          result = storage.ready();
+          expect(result).to.be.a('promise');
+          result.catch(err => {
+            expect(err).to.equal(evtErr);
+            expect(err).to.equal(error);
+            done();
+          });
+        });
+      });
+    });
+
+    afterEach(() => cleanStorage(storage));
+  });
+
+  afterEach(() => sinon.restore());
 });