+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.2.0
+
+ * Feature: Support for Mongoose connections
+ * Feature: Ready method to wait for the MongoDb connection
+ * Breaking change: Deprecated "connectionOpts" in favor of "options"
+
 # 3.1.0
 
  * Feature: Added caching feature